# order-service/docs/swagger.json
{
  "swagger": "2.0",
  "basePath": "/api/v1",
  "host": "localhost:3000",
  "info": {
    "title": "Order Service API",
    "description": "API documentation for Order Service",
    "version": "1.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "X-API-Key",
      "in": "header"
    }
  },
  "paths": {
...
    "/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get orders for a user",
        "description": "Returns paginated list of orders for the specified user ID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "description": "User ID (defaults to authenticated user)",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (defaults to 1)",
            "type": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page (defaults to 10, max 100)",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/model.OrderResponse"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Create a new order",
        "description": "Create a new order with items",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "order",
            "in": "body",
            "description": "Order creation request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/model.CreateOrderRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/model.OrderResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get order by ID",
        "description": "Returns order details for the specified ID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Order ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.OrderResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/orders/{id}/payment": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Process payment for an order",
        "description": "Process payment for a pending order",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Order ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/orders/{id}/status": {
      "patch": {
        "tags": [
          "Orders"
        ],
        "summary": "Update order status",
        "description": "Update the status of an order",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Order ID",
            "required": true,
            "type": "integer"
          },
          {
            "name": "status",
            "in": "body",
            "description": "New order status",
            "required": true,
            "schema": {
              "$ref": "#/definitions/model.UpdateOrderStatusRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
    "model.CreateOrderRequest": {
      "type": "object",
      "required": [
        "items",
        "payment_method",
        "shipping_address",
        "user_id"
      ],
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/model.OrderItemRequest"
          }
        },
        "payment_method": {
          "type": "string",
          "maxLength": 50
        },
        "shipping_address": {
          "type": "string"
        },
        "user_id": {
          "type": "string"
        }
      }
    },
...
    "model.OrderItemRequest": {
      "type": "object",
      "required": [
        "product_id",
        "quantity",
        "unit_price",
        "warehouse_id"
      ],
      "properties": {
        "order_id": {
          "type": "integer",
          "description": "Added for compatibility with warehouse service"
        },
        "product_id": {
          "type": "integer"
        },
        "quantity": {
          "type": "integer",
          "minimum": 1
        },
        "unit_price": {
          "type": "number",
          "minimum": 0
        },
        "warehouse_id": {
          "type": "integer"
        }
      }
    },
    "model.OrderItemResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "product_id": {
          "type": "integer"
        },
        "quantity": {
          "type": "integer"
        },
        "total_price": {
          "type": "number"
        },
        "unit_price": {
          "type": "number"
        },
        "warehouse_id": {
          "type": "integer"
        }
      }
    },
    "model.OrderResponse": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/model.OrderItemResponse"
          }
        },
        "payment_deadline": {
          "type": "string"
        },
        "payment_method": {
          "type": "string"
        },
        "shipping_address": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "total_amount": {
          "type": "number"
        },
        "updated_at": {
          "type": "string"
        },
        "user_id": {
          "type": "string"
        }
      }
    },
...
    "model.UpdateOrderStatusRequest": {
      "type": "object",
      "required": [
        "status"
      ],
      "properties": {
        "status": {
          "type": "string",
          "enum": [
            "pending",
            "paid",
            "cancelled",
            "completed"
          ]
        }
      }
    },
...
    "response.ErrorInfo": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "response.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/response.ErrorInfo"
        },
        "success": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "response.SuccessResponse": {
      "type": "object",
      "properties": {
        "data": {},
        "success": {
          "type": "boolean",
          "example": true
        }
      }
    }
  }
}